Added guard for null values on `put(key,value)`.
In the case of blank nodes being formed as WJSONObject instances, the
value being `put` for `@id` is `null`. This doesn't cause an immediate
failure, however later on when `toString()` on the WJSONObject is call a
NPE is thrown when `toString()` is applied, unguarded, to the null '@id'
value. This is caught (somewhere) and results in a `VelocityException`
and the failure of the invoked template.

The proposed fix here simply adds a guard in `put(key, value)` that
ignore a put of a null value. An alternate fix would be to alter
`WJSONObject.toString()` so that it checks for a null value prior to
expanding a `key` and its value and either suppress the null valued key
in the output or expand as a JSON NULL (if such a thing exists).